remove header files
