+++ puck.includes
@@ -1,9 +1,5 @@
 /opt/ros/melodic/include
 /usr/local/include/eigen3
 /usr/include/pcl-1.8
-~/Workspace/ws/loam_ws/devel/include
 src
-cfg
-logs/catkin_tools_prebuild
-.
-launch
+~/Workspace/ws/sv_ws/devel/include
